move PlainTextElement into its own file
